refactor(syncdeck): share embedded manager token parser
